refactor: 장바구니 컴포넌트 분리 및 공용 훅 추출
- CartButton을 cart 폴더로 이동하고 CartPopover 컴포넌트 분리
- 외부 클릭/ESC 감지 로직을 useClickOutside, useEscapeKey 훅으로 추출
- MobileNav에 useEscapeKey 적용
- 홈 페이지에 Header 적용
