+++ app/hooks/useEscapeKey.ts
@@ -1,0 +1,12 @@
+import {useEffect} from "react";
+
+export function useEscapeKey(onEscape: () => void, enabled: boolean) {
+  useEffect(() => {
+    if (!enabled) return;
+    const onKey = (e: KeyboardEvent) => {
+      if (e.key === "Escape") onEscape();
+    };
+    window.addEventListener("keydown", onKey);
+    return () => window.removeEventListener("keydown", onKey);
+  }, [enabled, onEscape]);
+}
